requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
pytz==2022.2.1
sqlparse==0.4.2
psycopg2-binary==2.8.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
tzdata==2022.4
djangorestframework-simplejwt==5.2.2